+++ clocks_stm32f103.h
@@ -15,7 +15,7 @@ void clk_config(enum CLKSRC source, bool pll_src_hse, bool pll_hse_div2, int pll
 
 void systick_count(int value);
 
-uint64_t get_hclock_frequency(void);
+uint32_t get_hclock_frequency(void);
 
 void crude_delay_cyc(uint64_t n); //Delay for n cycles using 1-cycle loop
 void crude_delay_ms(int n); //Delay for n milliseconds using crude_delay_cyc
